refactor package names for prediction methods.
