+++ transitclock/src/main/java/org/transitclock/core/predictiongenerator/scheduled/dwell/rls/TransitClockRLS.java
@@ -1,4 +1,4 @@
-package org.transitclock.core.predictiongenerator.rls.dwell.scheduled;
+package org.transitclock.core.predictiongenerator.scheduled.dwell.rls;
 
 import java.io.Serializable;
